text
this page used to say "next chapter," which was actually the core concept around which this project was built. it was incredibly useful to think of this page as a reoccurring element, something to come back to in between each chapter, something to ground it. i think the project has since outgrown its original idea, but it was crucial to the formation and i don't know if i would have been able to understand what i was writing without it. 

i have since taken to calling each portion a "page" instead of a chapter, so this phrasing has since become deprecated. it still makes sense to maintain a homepage (although there's little point until i can store user data and present readers with their correct, unique, consistent order) but if the homepage says "next chapter" that's gonna look like the title. 

working title changed from superscript to time.